working on Cache documentation
