+++ card/lib/card/cache.rb
@@ -2,13 +2,18 @@
 
 class Card
   class Cache
-    TEST_ENVS        = %w(test cucumber).freeze
-    @prepopulating   = TEST_ENVS.include? Rails.env
-    @no_rails_cache  = TEST_ENVS.include?(Rails.env) || ENV["NO_RAILS_CACHE"]
+    extend Card::Cache::Prepopulate
+
+    @prepopulating = %w(test cucumber).include? Rails.env
+    @no_rails_cache = %w(test cucumber).include?(Rails.env) ||
+                      ENV["NO_RAILS_CACHE"]
     @@cache_by_class = {}
     cattr_reader :cache_by_class
 
     class << self
+      # create a new cache for the ruby class provided
+      # @param klass [Class]
+      # @return [{Card::Cache}]
       def [] klass
         raise "nil klass" if klass.nil?
         cache_type = (@no_rails_cache ? nil : Cardio.cache)
@@ -26,11 +31,17 @@ class Card
         end
       end
 
-      def restore
+      # reset all caches for all classes
+      def reset_all
+        reset_hard
         reset_soft
-        prepopulate
+        reset_other
       end
 
+      # completely wipe out all caches, often including the Persistent cache of
+      # other decks using the same mechanism.
+      # Generally prefer {#reset_all}
+      # @see #reset_all
       def reset_global
         cache_by_class.each do |_klass, cache|
           cache.soft.reset
@@ -39,57 +50,39 @@ class Card
         reset_other
       end
 
-      def reset_all
-        reset_hard
-        reset_soft
-        reset_other
-      end
-
+      # reset the Persistent cache for all classes
       def reset_hard
         cache_by_class.each do |_klass, cache|
           cache.hard.reset if cache.hard
         end
       end
 
+      # reset the Temporary cache for all classes
       def reset_soft
         cache_by_class.each do |_klass, cache|
           cache.soft.reset
         end
       end
 
+      # reset Codename cache and delete tmp files
+      # (the non-standard caches)
       def reset_other
         Card::Codename.reset_cache
         Cardio.delete_tmp_files
       end
 
+      # generate a cache key from an object
+      # @param obj [Object]
+      # @return [String]
       def obj_to_key obj
         case obj
         when Hash
-          obj.sort.map do |key, value|
-            "#{key}=>(#{obj_to_key(value)})"
-          end.join ","
+          obj.sort.map { |key, value| "#{key}=>(#{obj_to_key(value)})" } * ","
         when Array
-          obj.map do |value|
-            obj_to_key(value)
-          end.join ","
+          obj.map { |value| obj_to_key(value) }
         else
           obj.to_s
         end
-      end
-
-      private
-
-      def prepopulate
-        return unless @prepopulating
-        soft = Card.cache.soft
-        @rule_cache ||= Card.rule_cache
-        @user_ids_cache ||= Card.user_ids_cache
-        @read_rule_cache ||= Card.read_rule_cache
-        @rule_keys_cache ||= Card.rule_keys_cache
-        soft.write "RULES", @rule_cache
-        soft.write "READRULES", @read_rule_cache
-        soft.write "USER_IDS", @user_ids_cache
-        soft.write "RULE_KEYS", @rule_keys_cache
       end
     end
 
@@ -98,9 +91,6 @@ class Card
     def initialize opts={}
       @klass = opts[:class]
       cache_by_class[@klass] = self
-
-      # hard cache mirrors the db
-      # only difference to db: it caches virtual cards
       @hard = Persistent.new opts if opts[:store]
 
       # soft cache is temporary
@@ -120,11 +110,7 @@ class Card
 
     def fetch key, &block
       @soft.fetch(key) do
-        if @hard
-          @hard.fetch(key, &block)
-        else
-          yield
-        end
+        @hard ? @hard.fetch(key, &block) : yield
       end
     end
 
@@ -149,13 +135,13 @@ class Card
   end
 end
 
-ActiveSupport::Cache::FileStore.class_eval do
-  # escape special symbols \*"<>| additionaly to :?.
-  # All of them not allowed to use in ms windows file system
-  def real_file_path name
-    name = name.gsub("%", "%25").gsub("?", "%3F").gsub(":", "%3A")
-    name = name.gsub('\\', "%5C").gsub("*", "%2A").gsub('"', "%22")
-    name = name.gsub("<", "%3C").gsub(">", "%3E").gsub("|", "%7C")
-    "%s/%s.cache" % [@cache_path, name]
-  end
-end
+# ActiveSupport::Cache::FileStore.class_eval do
+#   # escape special symbols \*"<>| additionaly to :?.
+#   # All of them not allowed to use in ms windows file system
+#   def real_file_path name
+#     name = name.gsub("%", "%25").gsub("?", "%3F").gsub(":", "%3A")
+#     name = name.gsub('\\', "%5C").gsub("*", "%2A").gsub('"', "%22")
+#     name = name.gsub("<", "%3C").gsub(">", "%3E").gsub("|", "%7C")
+#     "%s/%s.cache" % [@cache_path, name]
+#   end
+# end
